requirements.txt
boto3
transformers
torch
pillow
requests